refactor: 💡 use nest to run swagger generation instead of jest
