+++ .gitignore
@@ -1,6 +1,7 @@
 # compiled output
 /dist
 /node_modules
+polymesh-rest-api-swagger-spec.json
 
 # Logs
 logs
