+++ package.json
@@ -15,7 +15,8 @@
     "commit": "npx git-cz",
     "format": "prettier-eslint --write \"src/**/*.ts\" \"test/**/*.ts\"",
     "start": "nest start",
-    "start:repl": "yarn start --entryFile repl",
+    "start:repl": "yarn start --entryFile commands/repl",
+    "write-swagger": "yarn start --entryFile commands/write-swagger",
     "start:dev": "nest start --watch",
     "start:debug": "nest start --debug --watch",
     "start:prod": "NODE_ENV='production' node dist/main",
