+++ release.config.js
@@ -14,15 +14,9 @@ module.exports = {
     '@semantic-release/release-notes-generator',
     '@semantic-release/changelog',
     [
-      '@semantic-release/git',
-      {
-        assets: ['package.json', 'CHANGELOG.md', 'swagger-spec.json'],
-      },
-    ],
-    [
       '@semantic-release/github',
       {
-        assets: ['CHANGELOG.md', 'swagger-spec.json'],
+        assets: ['CHANGELOG.md', 'polymesh-rest-api-swagger-spec.json'],
       },
     ],
   ],
